Extract port logic to its own file.
