+++ shared/services/config/consensus-common-config.go
@@ -48,19 +48,7 @@ type ConsensusCommonConfig struct {
 
 // Create a new ConsensusCommonParams struct
 func NewConsensusCommonConfig(cfg *RocketPoolConfig) *ConsensusCommonConfig {
-	rpcPortModes := []config.ParameterOption{{
-		Name:        "Closed",
-		Description: "Do not allow connections to the RPC port",
-		Value:       config.RPC_Closed,
-	}, {
-		Name:        "Open to Localhost",
-		Description: "Allow connections from this host only",
-		Value:       config.RPC_OpenLocalhost,
-	}, {
-		Name:        "Open to External hosts",
-		Description: "Allow connections from external hosts. This is safe if you're running your node on your local network. If you're a VPS user, this would expose your node to the internet and could make it vulnerable to MEV/tips theft",
-		Value:       config.RPC_OpenExternal,
-	}}
+	portModes := config.PortModes("Allow connections from external hosts. This is safe if you're running your node on your local network. If you're a VPS user, this would expose your node to the internet and could make it vulnerable to MEV/tips theft")
 
 	return &ConsensusCommonConfig{
 		Title: "Common Consensus Client Settings",
@@ -126,7 +114,7 @@ func NewConsensusCommonConfig(cfg *RocketPoolConfig) *ConsensusCommonConfig {
 			EnvironmentVariables: []string{"BN_OPEN_API_PORT"},
 			CanBeBlank:           false,
 			OverwriteOnUpgrade:   false,
-			Options:              rpcPortModes,
+			Options:              portModes,
 		},
 
 		DoppelgangerDetection: config.Parameter{
