+++ shared/services/config/prometheus-config.go
@@ -30,19 +30,7 @@ type PrometheusConfig struct {
 
 // Generates a new Prometheus config
 func NewPrometheusConfig(cfg *RocketPoolConfig) *PrometheusConfig {
-	rpcPortModes := []config.ParameterOption{{
-		Name:        "Closed",
-		Description: "Do not allow connections to the RPC port",
-		Value:       config.RPC_Closed,
-	}, {
-		Name:        "Open to Localhost",
-		Description: "Allow connections from this host only",
-		Value:       config.RPC_OpenLocalhost,
-	}, {
-		Name:        "Open to External hosts",
-		Description: "Allow connections from external hosts. This is safe if you're running your node on your local network. If you're a VPS user, this would expose your node to the internet",
-		Value:       config.RPC_OpenExternal,
-	}}
+	rpcPortModes := config.PortModes("")
 	return &PrometheusConfig{
 		Title: "Prometheus Settings",
